+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Features:
 * Scroll selected jobs into view
 * Save JSON files if config changed (#15)
 * Allow for other vital criteria (#10)
+* Add command input feature (#14)
 
 Bugfixes:
 * Be more clear about why connections do not succeed (#20)